Extract a utility method to refresh the repository data
Affects #26: Query dialog does not restore values
